Cancel Automatic Selection
User can cancel the current automatic selection if:

 1. A selection has been created.
 2. User click on a selection start offset that is different with the current automatic selection start offset.

Example case: User click on a word not to select that word but to start a selection from the beginning of that word. After the word has selected (automatically), users can put their caret at the beginning of that word and then can start their own selection correctly (because the automatic selection has been cancelled once for that offset).